backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.25",
    "fastapi>=0.128.0",
    "mangum>=0.20.0",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
